requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.1
python-dotenv==1.0.0
